Added geofencing for the coin markers
Currently this only means that a "Coin encountered"
toast pops up when the user walks within 25 meters of a coin.

As Anko is used for the above toast, it is now a dependancy
for the project.

The new file GeofenceTransitionsIntentService is crucial to
listening to geofence events in teh background.

Changes to gradle files are to accomodate Anko.