User Authorization with Google Firebase
Added new activity "LoginActivity" which runs
when the app first launches. Here the user can
enter their email and password and either log in
or create an account to play the game with.
This relies on and uses the google firebase backend
which has been set up for the game.