Further refactoring, added untracked files
The refactoring was to adhere to the Kotlin
style guide better in regards to using lambas
over single-instance objects in handlers,
variable naming, and "when"s instead of cascading if/else
statements.

The untracked files added are necessary for the current
implementation of the app (navigating with bottom nav bar,
logging in and out, etc.) and were erroneously not added
to a previous commit.