User wallets with Firestore, KDoc changes
Added coins being added to the user's "Wallet"
in the Firestore database upon collection, and
also checking for if a coin is already in the wallet
when adding its marker (so that users cannot
collect the same coins twice).

Also updated and added more documentation for
the source code.

Finally added the Constants.kt file which ought to have
been included in an earlier commit.