Geofencing -> Approx distance calculation
Replaced the use of Google's Geofence API to notice when the user
gets close to the coins with instead manually approximating the
distance to each coin upon location change. This is as the geofences
were too slow at notifying the app when the user got close to a coin;
and the battery savings that cause this aren't really motivated
in this example as we need to pull the user's location constantly
anyway.