Ancient coins bonus feature, slight refactoring
Ancient coins are rare spawns which may spawn every
half hour between 7.00 - 18.30 and are worth a
large sum of coins. If such a coin spawns, a notification
shall fire and the coin will be available on the map
the next time the user starts the app.
Right now the spawn chance is 1.0 for testing purposes,
but later this will be lowered considerably.

The refactoring was centred around using when/if/elvis expressions
for variable assignments.