Fixed Ancient Coins bug
The locally stored data regarding the downloaded map and its
download date was not being updated when the download
was carried out in the background, causing the app to
redownload it for every alarm going off. This has now been fixed.

Also carried out some very minor refactoring to use
proper constant flags for the locally stored data.