Added custom icons
Added custom made icons for coins, ancient coins, and the bank.
This should greatly improve the the user experience as
it allows the user to quickly make judgements regarding
which coins to pick up first.

Also added a custom-made app icon.

Removed icons and drawables which are now unused, such as
the default app logo and the previous drawables
used for the bank and ancient coins.