Message creation bug fix, added progress bars
The message creation was reliant on LocalDateTime,
which caused the app to crash when run on API < 26.
This has been changed to use Calendar, which does
unfortunately mean that it will not be localized at all.

Added progress bars to many async tasks such as signing
in, updating listviews, sending messages, etc. to make
for a better UX.