Experimental change: FABs to switch user mode
This is the start of an experimental branch where
the user will enter either a specific "coin pick up"
or "coin inspection" mode using floating action buttons.
Picking up a coin will thus be done by being in the right mode
and then clicking on the marker, rather than there being a
pop up collection button as there is currently.