Randomized Message sending and receiving test
As it was before the test would succeed even if sending
message failed as long as there was a message with the
text "Test" at the top of the inbox. I have updated this
so that the text should also include a matching pseudo-
random double, which makes the chance of such an event
happening close to 0.