Fatal bugfix, changed distance calculation
Fixed a bug which caused the app to crach when
switching to AccountFragment with a combo active
by canceling any current combo in onStop().

Changed the distance calculation from the method
sourced from SO to using the build in LatLng
method from MapboxSDK as it is no longer necessary
to do this calculation very often (only on marker click,
rather than on location change as it was before).