Fixed bug causing alarms to not be set up on Oreo
The bug was due to Android Oreo requiring broadcast receivers
to be registered to the application context. This has now been
added, so the alarms should be set up correctly even on
later APIs!

Also removed an example test which was accidentally added recently,
and extended the comments in the test to clarify what they
expect in order to run appropriately.