Fixed bug with checking if app is run for 1st time
Bug was due to the fact that the flag was set
on the users device in LoginActivity. As such if the
user launched the app for the first time, quit before
logging in, and then relaunched it they would not
see the dialog greeting them the first time
MainActivity was run.