Extended and renamed test, more documentation
Extended and renamed the test which banks a coin from the users wallet.
Added documentation for tests which previously lacked this.

Also added more NPE-safety in the activities which use async queries,
and changed the default bank text as the query is now not fired
until the user chooses a source mode.