Added test in response to latest bug
Added a test in response to the bug fixed in the latest
commit in order to make sure that the issue does not
reappear with refactoring or changes made in the future.
The test first deposits coins from the wallet and then from
the inbox and makes sure that doing the latter does not
overwrite anything from the former (e.g. the counter of
how many coins were deposited today).