Added final test, slight change to shared prefs
Added a final test which tests the coin messaging functionality.

Also made a slight change to MainActivity; storing the
cachedMap is now done on demand instead of onStop when
the download date has been finished. As such overriding
onStop was no longer needed, and a different method
has superseeded it instead.